// types.ts
export type UserType = {
     uid?: string;
     email?: string | null;
     name: string | null;
     image?: any;
} | null;

export type UserDataType = {
     name: string;
     image?: any;
};

export type AuthContextType = {
     user: UserType;
     setUser: Function;
     login: (
     email: string,
     password: string
     ) => Promise<{ success: boolean; msg?: string }>;
     register: (
     email: string,
     password: string,
     name: string
     ) => Promise<{ success: boolean; msg?: string }>;
     updateUserData: (userId: string) => Promise<void>;
};